Care, you were incrementing i in the j loop.
Also, especially when using for loops over the array, don't use <= in the comparison bit, as you'll get off by one errors (crashes) all the time.  for (int i=0; i<10; i++) looks like a loop that goes ten times, and it does, 0-9, which fits an array with ten elements.

Also, move the print statements until after the sort... and done!